backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
weaviate-client
python-dotenv
langchain
langchain-community
langchain-text-splitters